mnt/requirements.txt
Pillow==8.4.0
numpy==1.20.1
keras==2.1.2
tensorflow-gpu==1.8.0